requirements.txt
uwsgi
sqlalchemy
flask_sqlalchemy
flask_migrate
pymysql
mysqlclient
